requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pycasbin==2.2.0
pydantic==2.11.7
pydantic-core==2.33.2
simpleeval==1.0.3
sniffio==1.3.1
starlette==0.47.3
typing-extensions==4.15.0
typing-inspection==0.4.1
uvicorn==0.35.0
